server/questions/requirements.txt
pdfplumber>=0.9.0
python-docx>=0.8.11
google-generativeai>=0.3.0
